# app/schemas/generated.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-10-04T19:31:42+00:00

from __future__ import annotations

from enum import StrEnum
from uuid import UUID

from pydantic import AwareDatetime, BaseModel, EmailStr, Field, conint, constr


class User(BaseModel):
    id: UUID
    name: str = Field(..., examples=['Alex'])
    email: EmailStr = Field(..., examples=['alex@example.com'])
    createdAt: AwareDatetime
    updatedAt: AwareDatetime


class LoginRequest(BaseModel):
    email: EmailStr
    password: str


class RegisterRequest(BaseModel):
    name: constr(min_length=1)
    email: EmailStr
    password: constr(min_length=8)


class Data(BaseModel):
    token: str
    user: User


class AuthResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: Data


class UserResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: User


class UpdateProfileRequest(BaseModel):
    name: str
    email: EmailStr


class ChangePasswordRequest(BaseModel):
    currentPassword: str
    newPassword: constr(min_length=8)


class PasswordResetRequest(BaseModel):
    email: EmailStr


class PasswordResetConfirm(BaseModel):
    token: str
    newPassword: constr(min_length=8)


class NotificationSettings(BaseModel):
    onScanCompletion: bool = Field(..., examples=[True])
    onScanFailure: bool = Field(..., examples=[True])


class NotificationSettingsResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: NotificationSettings


class Repository(BaseModel):
    id: UUID
    url: str = Field(..., examples=['https://github.com/org/repo'])
    name: str = Field(..., examples=['repo'])
    owner: str = Field(..., examples=['org'])
    provider: str = Field(..., examples=['github'])
    defaultBranch: str = Field(..., examples=['main'])
    isValid: bool
    addedBy: UUID
    createdAt: AwareDatetime
    updatedAt: AwareDatetime


class RepositoryResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: Repository


class ValidateRepoRequest(BaseModel):
    url: str


class Data1(BaseModel):
    valid: bool
    name: str | None = None
    owner: str | None = None
    defaultBranch: str | None = None


class ValidateRepoResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: Data1


class Branch(BaseModel):
    name: str = Field(..., examples=['main'])
    isDefault: bool
    lastCommit: str | None = Field(None, examples=['7f3b4c1'])


class BranchListResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: list[Branch]


class AddRepoRequest(BaseModel):
    url: str


class ScanStatus(StrEnum):
    queued = 'queued'
    running = 'running'
    completed = 'completed'
    failed = 'failed'
    cancelled = 'cancelled'


class FindingsCount(BaseModel):
    critical: int | None = Field(None, examples=[1])
    high: int | None = Field(None, examples=[3])
    medium: int | None = Field(None, examples=[5])
    low: int | None = Field(None, examples=[2])
    info: int | None = Field(None, examples=[0])


class Scan(BaseModel):
    id: UUID
    repositoryId: UUID
    branch: str = Field(..., examples=['main'])
    status: ScanStatus
    progress: conint(ge=0, le=100)
    findingsCount: FindingsCount
    startedAt: AwareDatetime | None = None
    completedAt: AwareDatetime | None = None
    cancelledAt: AwareDatetime | None = None
    createdAt: AwareDatetime
    updatedAt: AwareDatetime


class ScanResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: Scan


class Data2(BaseModel):
    id: UUID
    status: ScanStatus
    progress: int


class ScanProgressResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: Data2


class StartScanRequest(BaseModel):
    repositoryId: UUID
    branch: str = Field(..., examples=['develop'])


class Severity(StrEnum):
    critical = 'critical'
    high = 'high'
    medium = 'medium'
    low = 'low'
    info = 'info'


class ReviewStatus(StrEnum):
    open = 'open'
    acknowledged = 'acknowledged'
    false_positive = 'false_positive'
    resolved = 'resolved'


class Finding(BaseModel):
    id: UUID
    scanId: UUID
    repositoryId: UUID
    severity: Severity
    category: str = Field(..., examples=['Secret Exposure'])
    title: str = Field(..., examples=['Hardcoded AWS Key'])
    description: str
    filePath: str = Field(..., examples=['src/config.ts'])
    lineStart: int | None = None
    lineEnd: int | None = None
    codeSnippet: str | None = None
    recommendation: str
    aiExplanation: str | None = None
    reviewStatus: ReviewStatus
    reviewNote: str | None = None
    reviewedBy: UUID | None = None
    reviewedAt: AwareDatetime | None = None
    createdAt: AwareDatetime


class FindingResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: Finding


class ReviewFindingRequest(BaseModel):
    reviewStatus: ReviewStatus
    reviewNote: str | None = None


class ReportStatus(StrEnum):
    generating = 'generating'
    ready = 'ready'
    failed = 'failed'


class Report(BaseModel):
    id: UUID
    scanId: UUID
    status: ReportStatus
    format: str = Field(..., examples=['pdf'])
    fileUrl: str | None = None
    fileSize: int | None = None
    createdAt: AwareDatetime
    completedAt: AwareDatetime | None = None


class ReportResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: Report


class GenerateReportRequest(BaseModel):
    scanId: UUID


class Repositories(BaseModel):
    total: int | None = None


class Scans(BaseModel):
    total: int | None = None
    queued: int | None = None
    running: int | None = None
    completed: int | None = None
    failed: int | None = None
    cancelled: int | None = None


class Findings(BaseModel):
    total: int | None = None
    open: int | None = None
    critical: int | None = None
    high: int | None = None
    medium: int | None = None
    low: int | None = None
    info: int | None = None


class DashboardMetrics(BaseModel):
    repositories: Repositories
    scans: Scans
    findings: Findings
    recentScans: list[Scan]
    recentFindings: list[Finding]


class DashboardResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: DashboardMetrics


class GitHubIntegration(BaseModel):
    connected: bool
    username: str | None = None
    connectedAt: AwareDatetime | None = None


class GitHubResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: GitHubIntegration


class Pagination(BaseModel):
    page: int = Field(..., examples=[1])
    pageSize: int = Field(..., examples=[20])
    totalItems: int = Field(..., examples=[84])
    totalPages: int = Field(..., examples=[5])


class Data3(BaseModel):
    message: str | None = None


class GenericSuccess(BaseModel):
    status: str = Field(..., examples=['success'])
    data: Data3


class Detail(BaseModel):
    field: str | None = None
    message: str | None = None


class Error(BaseModel):
    code: str = Field(..., examples=['VALIDATION_ERROR'])
    message: str
    details: list[Detail] | None = None


class ErrorResponse(BaseModel):
    status: str = Field(..., examples=['error'])
    error: Error


class RepositoryListResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: list[Repository]
    pagination: Pagination


class ScanListResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: list[Scan]
    pagination: Pagination


class FindingListResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: list[Finding]
    pagination: Pagination


class ReportListResponse(BaseModel):
    status: str = Field(..., examples=['success'])
    data: list[Report]
    pagination: Pagination