# openapi.yaml
openapi: 3.0.3
info:
  title: Repo Security Auditor API
  version: 1.0.0
  description: REST API for repository security auditing, scanning, finding review, and PDF report generation.
servers:
  - url: http://localhost:8000/api
    description: Local Development
  - url: https://api.reposecurityauditor.io/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required: [id, name, email, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Alex
        email:
          type: string
          format: email
          example: alex@example.com
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UserResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/User'
    UpdateProfileRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
...
    NotificationSettings:
      type: object
      required: [onScanCompletion, onScanFailure]
      properties:
        onScanCompletion:
          type: boolean
          example: true
        onScanFailure:
          type: boolean
          example: true
    NotificationSettingsResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/NotificationSettings'
...
    ScanStatus:
      type: string
      enum: [queued, running, completed, failed, cancelled]
    FindingsCount:
      type: object
      properties:
        critical:
          type: integer
          example: 1
        high:
          type: integer
          example: 3
        medium:
          type: integer
          example: 5
        low:
          type: integer
          example: 2
        info:
          type: integer
          example: 0
    Scan:
      type: object
      required: [id, repositoryId, branch, status, progress, findingsCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        repositoryId:
          type: string
          format: uuid
        branch:
          type: string
          example: main
        status:
          $ref: '#/components/schemas/ScanStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        findingsCount:
          $ref: '#/components/schemas/FindingsCount'
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ScanResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/Scan'
    ScanListResponse:
      type: object
      required: [status, data, pagination]
      properties:
        status:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: '#/components/schemas/Scan'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ScanProgressResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          required: [id, status, progress]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/ScanStatus'
            progress:
              type: integer
    StartScanRequest:
      type: object
      required: [repositoryId, branch]
      properties:
        repositoryId:
          type: string
          format: uuid
        branch:
          type: string
          example: develop
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
    ReviewStatus:
      type: string
      enum: [open, acknowledged, false_positive, resolved]
    Finding:
      type: object
      required: [id, scanId, repositoryId, severity, category, title, description, filePath, recommendation, reviewStatus, createdAt]
      properties:
        id:
          type: string
          format: uuid
        scanId:
          type: string
          format: uuid
        repositoryId:
          type: string
          format: uuid
        severity:
          $ref: '#/components/schemas/Severity'
        category:
          type: string
          example: Secret Exposure
        title:
          type: string
          example: Hardcoded AWS Key
        description:
          type: string
        filePath:
          type: string
          example: src/config.ts
        lineStart:
          type: integer
          nullable: true
        lineEnd:
          type: integer
          nullable: true
        codeSnippet:
          type: string
          nullable: true
        recommendation:
          type: string
        aiExplanation:
          type: string
          nullable: true
        reviewStatus:
          $ref: '#/components/schemas/ReviewStatus'
        reviewNote:
          type: string
          nullable: true
        reviewedBy:
          type: string
          format: uuid
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    FindingListResponse:
      type: object
      required: [status, data, pagination]
      properties:
        status:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: '#/components/schemas/Finding'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    GitHubIntegration:
      type: object
      required: [connected]
      properties:
        connected:
          type: boolean
        username:
          type: string
          nullable: true
        connectedAt:
          type: string
          format: date-time
          nullable: true
    GitHubResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/GitHubIntegration'
    Pagination:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 84
        totalPages:
          type: integer
          example: 5
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      tags: [Users]
      summary: Update profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/me/password:
    put:
      tags: [Users]
      summary: Change password
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          description: Current password incorrect
  /users/me/settings:
    get:
      tags: [Users]
      summary: Read notification settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notification choices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSettingsResponse'
    put:
      tags: [Users]
      summary: Save notification settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettings'
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSettingsResponse'
...
  /scans:
    get:
      tags: [Scans]
      summary: Scan history
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: repositoryId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paged scans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanListResponse'
    post:
      tags: [Scans]
      summary: Start scan
      description: Also used for Rescan of failed, cancelled, or completed scans.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartScanRequest'
      responses:
        '201':
          description: Scan queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
  /scans/{id}:
    get:
      tags: [Scans]
      summary: Scan details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Scan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
  /scans/{id}/status:
    get:
      tags: [Scans]
      summary: Scan progress
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Status and progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanProgressResponse'
  /scans/{id}/cancel:
    post:
      tags: [Scans]
      summary: Cancel scan
      description: Only valid for queued or running scans.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Scan cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
        '422':
          description: Scan not in cancellable state
  /scans/{id}/findings:
    get:
      tags: [Scans]
      summary: Scan findings
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          schema:
            type: string
        - name: reviewStatus
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paged findings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingListResponse'
...
  /integrations/github:
    get:
      tags: [Integrations]
      summary: GitHub status
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Connection status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubResponse'
    post:
      tags: [Integrations]
      summary: Connect GitHub
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubResponse'
    delete:
      tags: [Integrations]
      summary: Disconnect GitHub
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Disconnected